feat: 完善平台功能 - ECharts修复、异常数据过滤、管理员CRUD、预警规则、操控记录等
- 修复 ECharts Vue3 Proxy 兼容性问题（图表实例移至非响应式属性）
- 新增异常传感器数据过滤（全零数据不再自动保存）
- 完善管理员后台：用户CRUD、环境数据管理、批量删除
- 新增操控历史记录模块（ControlHistory）
- 新增异常配置管理（ExceptionConfig）
- 完善预警规则 CRUD 和预警日志批量处理
- 优化设备控制和状态监控
- 前端 UI 优化：导航栏、登录注册页面、个人信息页
- 更新 .gitignore 排除根目录 node_modules

+++ idea/agri-backend/src/main/java/com/example/agribackend/controller/RealTimeController.java
@@ -34,7 +34,7 @@ public class RealTimeController {
             if (data != null && !data.isEmpty()) {
                 EnvDataDTO dto = mapToDto(data);
                 dto.setDataSource("device"); // 标记数据来源为设备
-                dto.setIsRealTime(true);     // 标记为实时数据
+                dto.setIsRealTime(true); // 标记为实时数据
                 logger.debug("成功获取ESP32数据: {}", dto);
                 return Result.success(Collections.singletonList(dto));
             }
@@ -75,7 +75,7 @@ public class RealTimeController {
                     soilAdcValue = 4095.0;
 
                 // 返回原始ADC值
-                dto.setSoilAdc(soilAdcValue);
+                dto.setSoilAdc(soilAdcValue.intValue());
                 // 同时计算百分比湿度作为参考
                 double soilMoisture = Math.max(0, Math.min(100, (4095 - soilAdcValue) / 40.95));
                 dto.setSoilMoisture(soilMoisture);
